+++ packages/webblocks-cms/config/webblocks-plugins.php
@@ -2,16 +2,9 @@
 
 return [
     'enabled' => [
-        'webblocks-ui-manager' => env('WEBBLOCKS_UI_MANAGER_ENABLED', false),
     ],
 
-    'webblocks_ui_manager' => [
-        'cdn_base_path' => env('WEBBLOCKS_UI_MANAGER_CDN_BASE_PATH', 'cdn/webblocks-ui'),
-        'cdn_base_url' => env('WEBBLOCKS_UI_MANAGER_CDN_BASE_URL'),
-        'expected_dist_files' => [
-            'webblocks-ui.css',
-            'webblocks-icons.css',
-            'webblocks-ui.js',
-        ],
+    'install' => [
+        'root' => env('WEBBLOCKS_PLUGIN_INSTALL_ROOT'),
     ],
 ];